requirements.txt
Django==5.1.3
celery==5.4.0
redis==5.0.8
twilio==9.3.0
django-filter==24.3
python-dotenv==1.0.1